# encodings/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Encodings Service
  description: Encoding example service demonstrating the use of different content
    types
  version: ""
host: localhost:80
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /concatbytes/{a}/{b}:
    get:
      tags:
      - text
      summary: concatbytes text
      operationId: text#concatbytes
      produces:
      - text/html
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: string
      - name: b
        in: path
        description: Right operand
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
            format: byte
      schemes:
      - http
  /concatbytesfield/{a}/{b}:
    get:
      tags:
      - text
      summary: concatbytesfield text
      operationId: text#concatbytesfield
      produces:
      - text/html
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: string
      - name: b
        in: path
        description: Right operand
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
            format: byte
      schemes:
      - http
  /concatstringfield/{a}/{b}:
    get:
      tags:
      - text
      summary: concatstringfield text
      operationId: text#concatstringfield
      produces:
      - text/html
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: string
      - name: b
        in: path
        description: Right operand
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /concatstrings/{a}/{b}:
    get:
      tags:
      - text
      summary: concatstrings text
      operationId: text#concatstrings
      produces:
      - text/html
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: string
      - name: b
        in: path
        description: Right operand
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
  /convert:
    post:
      tags:
      - xml
      summary: convert xml
      operationId: xml#convert
      parameters:
      - name: ConvertRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/XMLConvertRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/XMLConvertResponseBody'
      schemes:
      - http
  /swagger.json:
    get:
      summary: Download ../../gen/http/openapi.json
      operationId: text#/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
definitions:
  XMLConvertRequestBody:
    title: XMLConvertRequestBody
    type: object
    properties:
      description:
        type: string
        description: The description
        example: Debitis cum numquam laboriosam omnis omnis.
      name:
        type: string
        description: The name
        example: Minus provident asperiores accusantium nesciunt ut tempore.
    example:
      description: Rem minus voluptas at voluptates pariatur aut.
      name: In corrupti reiciendis excepturi eos.
  XMLConvertResponseBody:
    title: XMLConvertResponseBody
    type: object
    properties:
      description:
        type: string
        example: Laborum quis sed iusto sapiente nulla.
      name:
        type: string
        example: Asperiores laboriosam dolor.
    example:
      description: Sed ipsam aut dolorem voluptatem ipsa eaque.
      name: Accusamus quasi.